This fundamentally restructures Scrolls::Log
We move to a class from a module and on instantiation also
instantiate a class for our global_context that we freeze in
place. This removes the idea of adding to our global_context
once we've instantiated our logger.

The rewrite was to address the new signal/thread handling in
Ruby 2.0. Details can be found in
https://github.com/asenchi/scrolls/issues/53.

I think this also provides us with some flexibility around global
context in the future. Unfortunately this is not backward
compatiable with current Scrolls releases.