Introduce abstraction for stream logging
This introduces a wrapper class to put the logic for logging to an IO
like object. This is to prevent having to put all the logic in the
log.rb class itself based if it's a syslog instance or not.

It updates the tests to account for this and also removes the now unused
mutex code.