Fix merging log data on nested exception
This logic was refactored and before the return value of the .merge call
was directly used as an argument for the `log()` call. But since that
refactor, it was moved to a separate statement but it means that merging
it needs to use `merge!`.

Without `merge!` the newly returned hash isn't assigned to any variable
and lost, so we lose this in the logging context.